+++ server/src/ui/clientOptions.js
@@ -99,6 +99,16 @@ function clientOptions(perms) {
         perm: ['quests'],
         category: 'markers',
       },
+      showArQuestDotBadge: {
+        type: 'bool',
+        perm: ['quests'],
+        category: 'markers',
+      },
+      showNoArQuestDotBadge: {
+        type: 'bool',
+        perm: ['quests'],
+        category: 'markers',
+      },
       showArBadge: { type: 'bool', perm: ['pokestops'], category: 'markers' },
       invasionOpacity: {
         type: 'bool',